pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photometa-restore"
version = "0.1.0"
description = "A tool to restore metadata to photos from Google Photos Takeout"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "PhotoMeta Restore Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Utilities",
]
dependencies = [
    "Pillow>=10.0.0",
    "PySimpleGUI>=4.60.0",
    "piexif>=1.1.3",
    "win32-setctime>=1.1.0",
    "exif>=1.6.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "pathlib>=1.0.1",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/axatjpr/photometa-restore"
Issues = "https://github.com/axatjpr/photometa-restore/issues"

[project.scripts]
photometa-restore = "photometa_restore.__main__:main"

[tool.setuptools]
packages = ["photometa_restore"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100 